Revert "8458-l2-gate-warehouse-macro (#6631)"
This reverts commit 2389e302ad0466c01f82d675d080fd9b8794599b.

reversion is temporary to delay prod deploy
