+++ .github/workflows/rails_tests.yml
@@ -309,15 +309,6 @@ jobs:
         run: |
           OKTA_DOMAIN=localhost OKTA_CLIENT_ID=x OKTA_CLIENT_SECRET=x bundle exec rspec --fail-fast=$MAX_FAILURES --color -fd spec/requests/omniauth_spec.rb spec/requests/sessions_controller_spec.rb
 
-      - name: Run AUTH_METHOD=jwt tests
-        if: matrix.test_group.jwt
-        env:
-          MAX_FAILURES: 60
-          AWS_ACCESS_KEY_ID: local_access_key
-          AWS_SECRET_ACCESS_KEY: local_secret_key
-        run: |
-          AUTH_METHOD=jwt IDP_AUD=test-aud ISS_URL=https://idp.example.test JWKS_URL=https://idp.example.test/jwks JWT_ALGORITHM=RS256 bundle exec rspec --fail-fast=$MAX_FAILURES --color -fd spec/models/user_auth_method_spec.rb drivers/hmis/spec/models/hmis/user_auth_method_spec.rb spec/models/idp/jwt_helper_spec.rb spec/models/concerns/idp/jwt_user_spec.rb spec/controllers/concerns/idp/current_user_spec.rb
-
       - name: Run logging tests
         if: matrix.test_group.logging || needs.determine_matrix.outputs.logging == 'true'
         env:
