+++ drivers/hmis/app/models/hmis/user.rb
@@ -19,6 +19,7 @@ class Hmis::User < ApplicationRecord
   include Memery
 
   include UserConcern
+  include Idp::JwtUser
   include HasRecentItems
   include DeviseUserPatch
